Added three files
airline.csv
======
This file houses the data.  A portion of this data serves as the training set.  The remainder of the data serves as the testing set.

cs_airline.py
=======
This file is the Python code.  It
(a) loads airline.csv
(b) splits the data into training and testing sets using the forecast horizon (36 months) as the basis for performing the split
(c) trains the LSTM model
(d) tests the trained LSTM model

capstone_1.pdf
=========
This file discusses the theory and the results.